backend/requirements.txt
fastapi
uvicorn
python-multipart
exifread
Pillow
opencv-python
numpy
faiss-cpu
sentence-transformers
pydantic
